Combined the create and all the update dialogs into a single instance of a single component
Learned more about method handling and binding to state

+++ src/main/js/domain/product.js
@@ -5,6 +5,7 @@ import $ from 'jquery';
 const React = require('react');
 const ReactDOM = require('react-dom');
 
+
 function getId(object) {
 	let leaf = object;
 
@@ -36,6 +37,9 @@ export class ProductList extends React.Component {
 	constructor(props) {
 		super(props);
 		this.handleInput = this.handleInput.bind(this);
+		this.handlePrepareUpdate = this.handlePrepareUpdate.bind(this);
+		this.handlePrepareCreate = this.handlePrepareCreate.bind(this);
+		this.editProductRef = React.createRef();
 	}
 
 	handleInput(e) {
@@ -48,6 +52,15 @@ export class ProductList extends React.Component {
 		}
 	}
 
+	handlePrepareUpdate(product) {
+		let editDialog = this.editProductRef.current;
+		editDialog.prepareUpdate(product);
+	}
+
+	handlePrepareCreate() {
+		let editDialog = this.editProductRef.current;
+		editDialog.prepareCreate();
+	}
 
 
 	render() {
@@ -57,10 +70,9 @@ export class ProductList extends React.Component {
 									attributes={this.props.attributes}
 									onDelete={this.props.onDelete}
 									onUpdate={this.props.onUpdate}
+									handlePrepareUpdate={this.handlePrepareUpdate}
 			/>
 		);
-
-		const createDialog = "createProduct";
 
 		return (
 			<div className="container">
@@ -81,13 +93,16 @@ export class ProductList extends React.Component {
 					</tbody>
 				</table>
 
-				<button type="button" className="btn btn-primary" data-toggle="modal" data-target={"#" + createDialog}>
+				<button type="button" className="btn btn-primary" onClick={this.handlePrepareCreate}>
 					<i className="fa fa-plus" /> Create Product
 				</button>
 
-				<CreateProductDialog attributes={this.props.attributes}
-														 onCreate={this.props.onCreate}
-														 dialogId={createDialog}
+				<EditProductDialog ref={this.editProductRef}
+													 attributes={this.props.attributes}
+													 onCreate={this.props.onCreate}
+														 // dialogId={createDialog}
+													 onUpdate={this.props.onUpdate}
+													 product={null}
 				/>
 
 				<Paginator
@@ -104,16 +119,21 @@ export class ProductRow extends React.Component {
 	constructor(props) {
 		super(props);
 		this.handleDelete = this.handleDelete.bind(this);
+		this.showUpdate = this.showUpdate.bind(this);
 	}
 
 	handleDelete() {
-		this.props.onDelete(this.props.product);
+		if (confirm("Delete " + this.props.product.entity.name + "?")) {
+			this.props.onDelete(this.props.product);
+		}
+	}
+
+	showUpdate() {
+		this.props.handlePrepareUpdate(this.props.product);
 	}
 
 	render() {
 		let entity = this.props.product.entity;
-
-		const dialogId = "updateProduct-" + getId(entity);
 
 		return (
 			<tr>
@@ -127,16 +147,11 @@ export class ProductRow extends React.Component {
 							<i className="fa fa-trash-alt" /> Delete
 						</button>
 
-						<button type="button" className="btn btn-primary" data-toggle="modal" data-target={"#" + dialogId}>
+						{/*<button type="button" className="btn btn-primary" onClick={() => this.showUpdate()}>*/}
+						<button type="button" className="btn btn-primary" onClick={this.showUpdate}>
 							<i className="fa fa-edit" /> Update
 						</button>
 					</div>
-
-					<UpdateProductDialog product={this.props.product}
-															 attributes={this.props.attributes}
-															 onUpdate={this.props.onUpdate}
-															 dialogId={dialogId}
-												 />
 				</td>
 			</tr>
 		)
@@ -144,45 +159,78 @@ export class ProductRow extends React.Component {
 }
 
 
-export class CreateProductDialog extends React.Component {
+export class EditProductDialog extends React.Component {
 	constructor(props) {
 		super(props);
 		this.handleSubmit = this.handleSubmit.bind(this);
+		this.prepareUpdate = this.prepareUpdate.bind(this);
+		this.prepareCreate = this.prepareCreate.bind(this);
+
+		this.state = {product: null};
 	}
 
-	handleSubmit(e) {
-		e.preventDefault();
-
-		const newProduct = {};
-
+	prepareUpdate(product) {
+		this.setState({product: product});
 		this.props.attributes.forEach(attribute => {
-			newProduct[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
+			ReactDOM.findDOMNode(this.refs[attribute]).value = product.entity[attribute].trim();
 		});
-		this.props.onCreate(newProduct);
+
+		$('#productDialog').modal('show');
+	}
+
+	prepareCreate() {
+		this.setState({product: null});
 
 		// clear out the dialog's inputs
 		this.props.attributes.forEach(attribute => {
 			ReactDOM.findDOMNode(this.refs[attribute]).value = '';
 		});
 
-		// Navigate away from the dialog to hide it.
-		window.location = "#";
+		$('#productDialog').modal('show');
+	}
+
+	handleSubmit(e) {
+		e.preventDefault();
+
+		if (this.state.product) {
+			// do update
+			const product = {};
+
+			this.props.attributes.forEach(attribute => {
+				product[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
+			});
+			this.props.onUpdate(this.state.product, product);
+
+		} else {
+			// do create
+			const product = {};
+
+			this.props.attributes.forEach(attribute => {
+				product[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
+			});
+			this.props.onCreate(product);
+
+		}
+
+		$('#productDialog').modal('hide');
 	}
 
 	render() {
 		return (
 			<div>
-				<div id={this.props.dialogId} className="modal fade" tabIndex="-1" role="dialog">
+				<div id="productDialog" className="modal fade" tabIndex="-1" role="dialog">
+				{/*<div id={this.props.dialogId}>*/}
 					<div className="modal-dialog">
 						<div className="modal-content">
 							<div className="modal-header">
-								<h5 className="modal-title">Create product</h5>
+								<h5 className="modal-title">{this.state.product ? "Update" : "Create"} product</h5>
 								<button type="button" className="close" data-dismiss="modal">
 									<span>&times;</span>
 								</button>
 							</div>
 							<div className="modal-body">
 								<form>
+									{/*<input type="hidden" ref="id" />*/}
 									<div className="form-group row">
 										<label htmlFor="name" className="col-sm-2 col-form-label">Name</label>
 										<div className="col-sm-6">
@@ -214,7 +262,10 @@ export class CreateProductDialog extends React.Component {
 							</div>
 							<div className="modal-footer">
 								<button type="button" className="btn" data-dismiss="modal">Cancel</button>
-								<button type="button" className="btn btn-primary" onClick={this.handleSubmit}>Create</button>
+								<button type="button" className="btn btn-primary" onClick={this.handleSubmit}>
+									<i className="fa fa-save" />
+									{this.state.product ? " Update" : " Create"}
+								</button>
 							</div>
 						</div>
 					</div>
@@ -224,85 +275,83 @@ export class CreateProductDialog extends React.Component {
 	}
 }
 
-export class UpdateProductDialog extends React.Component {
-	constructor(props) {
-		super(props);
-		this.handleSubmit = this.handleSubmit.bind(this);
-	}
-
-	handleSubmit(e) {
-		e.preventDefault();
-
-		console.log("Updating ", this.props.product);
-
-		const updatedProduct = {};
-
-		this.props.attributes.forEach(attribute => {
-			updatedProduct[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
-		});
-		this.props.onUpdate(this.props.product, updatedProduct);
-
-		// Navigate away from the dialog to hide it.
-		// window.location = "#";
-		$('#' + this.props.dialogId).modal('hide');
-	}
-
-	render() {
-
-		const entity = this.props.product.entity;
-		// const dialogId = "updateProduct-" + getId(entity);
-
-		return (
-			<span>
-				<div id={this.props.dialogId} className="modal fade" tabIndex="-1" role="dialog">
-					<div className="modal-dialog">
-						<div className="modal-content">
-							<div className="modal-header">
-								<h5 className="modal-title">Update product</h5>
-								<button type="button" className="close" data-dismiss="modal">
-									<span>&times;</span>
-								</button>
-							</div>
-							<div className="modal-body">
-								<form>
-									<div className="form-group row">
-										<label htmlFor="name" className="col-sm-2 col-form-label">Name</label>
-										<div className="col-sm-6">
-											<input type="text" className="form-control" ref="name" id="name" defaultValue={entity.name}/>
-										</div>
-									</div>
-
-									<div className="form-group row">
-										<label htmlFor="unit" className="col-sm-2 col-form-label">Unit</label>
-										<div className="col-sm-3">
-											<select className="form-control" ref="measureUnit" id="unit" defaultValue={entity.measureUnit}>
-												<option></option>
-												<option value="GRAM">g</option>
-												<option value="KILOGRAM">kg</option>
-												<option value="LITRE">l</option>
-												<option value="COUNT">count</option>
-											</select>
-										</div>
-									</div>
-
-									<div className="form-group row">
-										<label htmlFor="category" className="col-sm-2 col-form-label">Category</label>
-										<div className="col-sm-3">
-											<input type="text" className="form-control" ref="category" id="category" defaultValue={entity.category}/>
-										</div>
-									</div>
-
-								</form>
-
-							</div>
-							<div className="modal-footer">
-								<button type="button" className="btn" data-dismiss="modal">Cancel</button>
-								<button type="button" className="btn btn-primary" onClick={this.handleSubmit}>Update</button>
-							</div>
-						</div>
-					</div>
-				</div>
-			</span>
-		)
-	}
-}
+// export class UpdateProductDialog extends React.Component {
+// 	constructor(props) {
+// 		super(props);
+// 		this.handleSubmit = this.handleSubmit.bind(this);
+// 	}
+//
+// 	handleSubmit(e) {
+// 		e.preventDefault();
+//
+// 		console.log("Updating ", this.props.product);
+//
+// 		const updatedProduct = {};
+//
+// 		this.props.attributes.forEach(attribute => {
+// 			updatedProduct[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
+// 		});
+// 		this.props.onUpdate(this.props.product, updatedProduct);
+//
+// 		$('#' + this.props.dialogId).modal('hide');
+// 	}
+//
+// 	render() {
+//
+// 		const entity = this.props.product.entity;
+//
+// 		return (
+// 			<span>
+// 				<div id={this.props.dialogId} className="modal fade" tabIndex="-1" role="dialog">
+// 				{/*<div id={this.props.dialogId}>*/}
+// 					<div className="modal-dialog">
+// 						<div className="modal-content">
+// 							<div className="modal-header">
+// 								<h5 className="modal-title">Update product</h5>
+// 								<button type="button" className="close" data-dismiss="modal">
+// 									<span>&times;</span>
+// 								</button>
+// 							</div>
+// 							<div className="modal-body">
+// 								<form>
+// 									<div className="form-group row">
+// 										<label htmlFor="name" className="col-sm-2 col-form-label">Name</label>
+// 										<div className="col-sm-6">
+// 											<input type="text" className="form-control" ref="name" id="name" defaultValue={entity.name}/>
+// 										</div>
+// 									</div>
+//
+// 									<div className="form-group row">
+// 										<label htmlFor="unit" className="col-sm-2 col-form-label">Unit</label>
+// 										<div className="col-sm-3">
+// 											<select className="form-control" ref="measureUnit" id="unit" defaultValue={entity.measureUnit}>
+// 												<option></option>
+// 												<option value="GRAM">g</option>
+// 												<option value="KILOGRAM">kg</option>
+// 												<option value="LITRE">l</option>
+// 												<option value="COUNT">count</option>
+// 											</select>
+// 										</div>
+// 									</div>
+//
+// 									<div className="form-group row">
+// 										<label htmlFor="category" className="col-sm-2 col-form-label">Category</label>
+// 										<div className="col-sm-3">
+// 											<input type="text" className="form-control" ref="category" id="category" defaultValue={entity.category}/>
+// 										</div>
+// 									</div>
+//
+// 								</form>
+//
+// 							</div>
+// 							<div className="modal-footer">
+// 								<button type="button" className="btn" data-dismiss="modal">Cancel</button>
+// 								<button type="button" className="btn btn-primary" onClick={this.handleSubmit}>Update</button>
+// 							</div>
+// 						</div>
+// 					</div>
+// 				</div>
+// 			</span>
+// 		)
+// 	}
+// }
